refactor(image-output): streamline image processing modules - Replace custom output handlers with `getImageOutputFiles` - Simplify result rendering with `ImageResults` component - Update dependencies for improved performance and maintainability - Remove redundant code for cleaning up file operations
