+++ extensions/delphi-tools-raycast-extension/docs/delphitools/split.md
@@ -37,5 +37,3 @@ Tile image files.
 - `Copy All Tile Paths`: copy a list of all generated tile paths.
 - `Reveal Output Folder`: reveal the temporary output directory.
 - `Reveal in Finder`: reveal the selected tile file in Finder.
-
-
