+++ extensions/delphi-tools-raycast-extension/src/rmbg.tsx
@@ -1,19 +1,21 @@
 import { execFileAsync, getDelphitoolsCliPath } from "./utils/exec";
-import { getDefaultOutputRoot } from "./utils/preferences";
+import {
+  createOutputDirectory,
+  getImageOutputFiles,
+  ImageOutput,
+  ImageOutputResult,
+  ImageResults,
+} from "./utils/image-output";
 import {
   Action,
   ActionPanel,
-  Clipboard,
   Form,
   Icon,
-  List,
   LocalStorage,
   showToast,
   Toast,
   useNavigation,
 } from "@raycast/api";
-import { mkdir, readdir, stat } from "node:fs/promises";
-import path from "node:path";
 import { useEffect, useState } from "react";
 
 import {
@@ -25,16 +27,8 @@ type FormValues = {
   images: string[];
 };
 
-type RmbgOutput = {
-  path: string;
-  size: number;
-};
-
-type RmbgResult = {
-  outputDirectory: string;
-  outputs: RmbgOutput[];
-  outputPaths: string[];
-};
+type RmbgOutput = ImageOutput;
+type RmbgResult = ImageOutputResult<RmbgOutput>;
 
 const OUTPUT_NAMESPACE = "rmbg";
 const FIRST_RUN_KEY = "rmbg-has-run";
@@ -133,95 +127,20 @@ function RmbgForm({ isCheckingInstall }: { isCheckingInstall: boolean }) {
 }
 
 function RmbgResults({ result }: { result: RmbgResult }) {
-  const allPaths = result.outputPaths.join("\n");
-
   return (
-    <List searchBarPlaceholder="Search processed images">
-      {result.outputs.map((output) => (
-        <List.Item
-          key={output.path}
-          icon={{ source: output.path }}
-          title={path.basename(output.path)}
-          subtitle={path.dirname(output.path)}
-          accessories={[{ text: formatFileSize(output.size) }]}
-          actions={
-            <RmbgActions
-              outputPath={output.path}
-              allPaths={allPaths}
-              outputDirectory={result.outputDirectory}
-            />
-          }
-        />
-      ))}
-    </List>
-  );
-}
-
-function RmbgActions({
-  outputPath,
-  allPaths,
-  outputDirectory,
-}: {
-  outputPath: string;
-  allPaths: string;
-  outputDirectory: string;
-}) {
-  async function copyImage() {
-    await Clipboard.copy({ file: outputPath });
-    await showToast({
-      style: Toast.Style.Success,
-      title: "Copied Image",
-    });
-  }
-
-  async function copyAllPaths() {
-    await Clipboard.copy(allPaths);
-    await showToast({
-      style: Toast.Style.Success,
-      title: "Copied All Output Paths",
-    });
-  }
-
-  return (
-    <ActionPanel>
-      <Action.Open icon={Icon.Eye} title="Open Image" target={outputPath} />
-      <Action icon={Icon.Clipboard} title="Copy Image" onAction={copyImage} />
-      <Action.CopyToClipboard
-        title="Copy Image Path"
-        content={outputPath}
-        shortcut={{ modifiers: ["cmd", "shift"], key: "c" }}
-      />
-      <Action
-        icon={Icon.Clipboard}
-        title="Copy All Output Paths"
-        shortcut={{ modifiers: ["cmd", "shift"], key: "a" }}
-        onAction={copyAllPaths}
-      />
-      <Action.ShowInFinder
-        title="Reveal Output Folder"
-        path={outputDirectory}
-        shortcut={{ modifiers: ["cmd", "shift"], key: "r" }}
-      />
-      <Action.ShowInFinder
-        title="Reveal in Finder"
-        path={outputPath}
-        shortcut={{ modifiers: ["cmd"], key: "r" }}
-      />
-    </ActionPanel>
+    <ImageResults
+      result={result}
+      searchBarPlaceholder="Search processed images"
+      openTitle="Open Image"
+      copyImageTitle="Copy Image"
+      copyImagePathTitle="Copy Image Path"
+      copiedImageTitle="Copied Image"
+    />
   );
 }
 
 async function runRmbg(images: string[]): Promise<RmbgResult> {
-  const outputRoot = getDefaultOutputRoot();
-  await mkdir(outputRoot, { recursive: true });
-
-  const hash = Math.random().toString(36).slice(2, 10);
-  const outputDirectory = path.join(
-    outputRoot,
-    OUTPUT_NAMESPACE,
-    `${Date.now()}-${hash}`,
-  );
-  await mkdir(outputDirectory, { recursive: true });
+  const outputDirectory = await createOutputDirectory(OUTPUT_NAMESPACE);
 
   const args = [
     "rmbg",
@@ -234,7 +153,7 @@ async function runRmbg(images: string[]): Promise<RmbgResult> {
 
   await execFileAsync(getDelphitoolsCliPath(), args);
 
-  const outputs = await getOutputFiles(outputDirectory);
+  const outputs = await getImageOutputFiles(outputDirectory);
 
   if (outputs.length === 0) {
     throw new Error("No processed images were generated.");
@@ -245,39 +164,4 @@ async function runRmbg(images: string[]): Promise<RmbgResult> {
     outputs,
     outputPaths: outputs.map((out) => out.path),
   };
-}
-
-async function getOutputFiles(outputDirectory: string): Promise<RmbgOutput[]> {
-  const entries = await readdir(outputDirectory);
-  const files = await Promise.all(
-    entries.map(async (entry) => {
-      const outputPath = path.join(outputDirectory, entry);
-      const outputStat = await stat(outputPath);
-
-      return outputStat.isFile()
-        ? { path: outputPath, size: outputStat.size }
-        : null;
-    }),
-  );
-
-  return files
-    .filter((file): file is RmbgOutput => file !== null)
-    .sort((a, b) => a.path.localeCompare(b.path));
-}
-
-function formatFileSize(bytes: number): string {
-  const units = ["B", "KB", "MB", "GB"];
-  let size = bytes;
-  let unitIndex = 0;
-
-  while (size >= 1024 && unitIndex < units.length - 1) {
-    size /= 1024;
-    unitIndex += 1;
-  }
-
-  if (unitIndex === 0) {
-    return `${bytes} B`;
-  }
-
-  return `${size >= 10 ? size.toFixed(0) : size.toFixed(1)} ${units[unitIndex]}`;
 }
