+++ extensions/delphi-tools-raycast-extension/tsconfig.json
@@ -4,7 +4,8 @@
   "compilerOptions": {
     "lib": ["ES2023"],
     "module": "commonjs",
-    "target": "ES2023",
+    "target": "ES2022",
+    "types": ["node"],
     "strict": true,
     "isolatedModules": true,
     "esModuleInterop": true,
